Indice redare culoare este float.

+++ src/main/java/ro/cti/ugal/proiect/java/SursaIluminat.java
@@ -4,20 +4,20 @@ public abstract class SursaIluminat {
     public int luminozitate;            // In lumens
     public int temperaturaCuloare;      // In Kelvin, culoarea luminii emisa de bec.
     public int consum;                  // In watts
-    public double indiceRedareCuloare;  // Procent - [0.0, 1.0]
+    public float indiceRedareCuloare;  // Procent - [0.0, 1.0]
     
     public SursaIluminat() {
         this.luminozitate = 0;
         this.temperaturaCuloare = 0;
         this.consum = 0;
-        this.indiceRedareCuloare = 0.0;
+        this.indiceRedareCuloare = 0.0f;
     }
     
-    public SursaIluminat(int luminozitate, int temperaturaCuloare, int consum, double indiceRedareCuloare) {
+    public SursaIluminat(int luminozitate, int temperaturaCuloare, int consum, float indiceRedareCuloare) {
         this.luminozitate = luminozitate;
         this.temperaturaCuloare = temperaturaCuloare;
         this.consum = consum;
-        this.indiceRedareCuloare = Math.clamp(indiceRedareCuloare, 0.0, 1.0);
+        this.indiceRedareCuloare = Math.clamp(indiceRedareCuloare, 0.0f, 1.0f);
     }
     
     public SursaIluminat(SursaIluminat o) {
